% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.list.errors}
\alias{as.list.errors}
\title{Coerce to a List}
\usage{
\method{as.list}{errors}(x, ...)
}
\arguments{
\item{x}{object to be coerced or tested.}

\item{...}{objects, possibly named.}
}
\description{
S3 method for \code{errors} objects (see \code{\link{as.list}}).
}
\examples{
x <- set_errors(1:3, 0.1)
as.list(x)

}
