% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as.matrix.errors}
\alias{as.matrix.errors}
\title{Coerce to a Matrix}
\usage{
\method{as.matrix}{errors}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
S3 method for \code{errors} objects (see \code{\link{as.matrix}}).
}
\examples{
as.matrix(set_errors(1:3, 0.1))

}
