% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{c.errors}
\alias{c.errors}
\title{Combine Values into a Vector or List}
\usage{
\method{c}{errors}(...)
}
\arguments{
\item{...}{objects to be concatenated.}
}
\description{
S3 method for \code{errors} objects (see \code{\link{c}}).
}
\examples{
c(set_errors(1, 0.2), set_errors(7:9, 0.1), 3)

}
