% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{covar}
\alias{covar}
\alias{covar}
\alias{covar<-}
\title{Set Correlations Between \code{errors} objects}
\usage{
covar(x, y)

covar(x, y) <- value
}
\arguments{
\item{x}{a numeric object, or object of class \code{errors}.}

\item{y}{an object of class \code{errors}.}

\item{value}{a numeric vector of length 1 or the same length as \code{x}.}
}
\description{
Set/retrieve correlations.
}
