% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{drop_errors}
\alias{drop_errors}
\title{Drop Errors}
\usage{
drop_errors(x)
}
\arguments{
\item{x}{an \code{errors} object.}
}
\value{
the numeric without any \code{errors} attributes, while preserving other
  attributes like dimensions or other classes.
}
\description{
Drop Errors
}
\note{
Equivalent to \code{errors(x) <- NULL} or \code{set_errors(x, NULL)}.
}
