% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{package}
\name{errors-package}
\alias{errors-package}
\title{\pkg{errors}: Uncertainty Propagation for R Vectors}
\description{
Every measurement has an unknown error associated. Uncertainty is the
acknowledgement of that error: we are aware that our representation of reality
may differ from reality itself. This package provides support for measurement
errors in \R vectors, matrices and arrays. Uncertainty metadata is associated
to quantity values, and this uncertainty is automatically propagated when you
operate with \code{errors} objects, or with \code{errors} and numeric objects
(then numeric values are automatically coerced to \code{errors} with no uncertainty).
}
\details{
This package treats uncertainty as coming from Gaussian and linear sources,
and propagates them using the first-order Taylor series method for
propagation of uncertainty. Although the above assumptions are valid in a wide
range of applications in science and engineering, the practitioner should
evaluate whether they apply for each particular case.
}
\author{
Iñaki Ucar
}
