% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{matmult.errors}
\alias{matmult.errors}
\alias{\%*\%}
\title{Matrix Multiplication}
\usage{
x \%*\% y
}
\arguments{
\item{x}{numeric or complex matrices or vectors.}

\item{y}{numeric or complex matrices or vectors.}
}
\description{
Not implemented for \code{errors} objects. Errors are dropped.
}
