% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{matmult}
\alias{matmult}
\alias{\%*\%}
\title{Matrix Multiplication}
\usage{
x \%*\% y
}
\arguments{
\item{x}{numeric matrix or vector}

\item{y}{numeric matrix or vector}
}
\description{
Matrix Multiplication
}
