% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.errors}
\alias{plot.errors}
\title{X-Y plotting with errorbars}
\usage{
\method{plot}{errors}(x, y, ...)
}
\arguments{
\item{x}{object of class errors to plot along the x axis, or, if y is missing, along the y axis.}

\item{y}{object of class errors to plot along the y axis, or missing.}

\item{...}{other parameters, passed on to \link{plot.default}.}
}
\description{
Method for plotting \code{errors} objects.
}
