% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.errors}
\alias{print.errors}
\title{Print Values}
\usage{
\method{print}{errors}(x, ...)
}
\arguments{
\item{x}{an \code{errors} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method for \code{errors} objects.
}
\examples{
x <- set_errors(1:10, 1:10 * 0.05)
print(x)
print(x[1:3])
print(x[1])
print(x[1], digits=2)
print(x[1], notation="plus-minus")

}
