% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_errors.R
\name{scale_errors}
\alias{scale_x_errors}
\alias{scale_y_errors}
\alias{scale_colour_errors}
\alias{scale_color_errors}
\alias{scale_fill_errors}
\alias{scale_alpha_errors}
\alias{scale_size_errors}
\alias{scale_size_area_errors}
\alias{scale_radius_errors}
\alias{scale_linewidth_errors}
\title{Continuous scales for \code{errors} objects}
\usage{
scale_x_errors(...)

scale_y_errors(...)

scale_colour_errors(...)

scale_color_errors(...)

scale_fill_errors(...)

scale_alpha_errors(...)

scale_size_errors(...)

scale_size_area_errors(...)

scale_radius_errors(...)

scale_linewidth_errors(...)
}
\arguments{
\item{...}{arguments passed on to the corresponding continuous scale
(see the manual page for each \code{scale_{type}} for details).}
}
\description{
Default scales for the \code{errors} class.
}
