% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{t.errors}
\alias{t.errors}
\title{Matrix Transpose}
\usage{
\method{t}{errors}(x)
}
\arguments{
\item{x}{a matrix or data frame, typically.}
}
\description{
S3 method for \code{errors} objects (see \code{\link{t}}).
}
\examples{
a <- matrix(1:30, 5, 6)
errors(a) <- 1:30
t(a)

}
