% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tibble}
\alias{tibble}
\alias{type_sum.errors}
\alias{pillar_shaft.errors}
\title{Methods for Tidy \code{tibble} Printing}
\usage{
type_sum.errors(x)

pillar_shaft.errors(x, ...)
}
\arguments{
\item{x}{object of class errors.}

\item{...}{see \link[pillar]{pillar_shaft}.}
}
\description{
S3 methods for \code{errors} objects.
}
