% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{type_sum.errors}
\alias{type_sum.errors}
\title{type_sum for tidy tibble printing}
\usage{
type_sum.errors(x, ...)
}
\arguments{
\item{x}{object of class errors}

\item{...}{ignored}
}
\description{
type_sum for tidy tibble printing
}
