% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{drop_errors}
\alias{drop_errors}
\alias{drop_errors.data.frame}
\title{Drop Uncertainty}
\usage{
drop_errors(x)

\method{drop_errors}{data.frame}(x)
}
\arguments{
\item{x}{an \code{errors} object.}
}
\value{
the numeric without any \code{errors} attributes, while preserving other
  attributes like dimensions or other classes.
}
\description{
Drop Uncertainty
}
\note{
Equivalent to \code{errors(x) <- NULL} or \code{set_errors(x, NULL)}.
}
