% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{package}
\name{errors-package}
\alias{errors-package}
\title{\pkg{errors}: Uncertainty Propagation for R Vectors}
\description{
Support for measurement errors in R vectors, matrices and arrays: automatic
uncertainty propagation and reporting.
Errors are automatically propagated when you operate with \code{errors}
objects, or with \code{errors} and numeric objects (then numeric values are
automatically coerced to errors with zero error).
}
\details{
This package treats errors as coming from Gaussian, linear and independent
sources, and propagates them using the first-order Taylor series method for
propagation of uncertainty. Although the above assumptions are valid in a wide
range of applications in science and engineering, the practitioner should
evaluate whether they apply for each particular case.
}
\author{
Iñaki Ucar
}
