% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{errors}
\alias{errors}
\alias{errors_max}
\alias{errors_min}
\alias{errors<-}
\alias{set_errors}
\alias{as.errors}
\title{Handle Uncertainty on a Numeric Vector}
\usage{
errors(x)

errors_max(x)

errors_min(x)

errors(x) <- value

set_errors(x, value = 0)

as.errors(x, value = 0)
}
\arguments{
\item{x}{a numeric object, or object of class \code{errors}.}

\item{value}{a numeric vector of length 1 or the same length as \code{x}.}
}
\value{
\code{errors} returns a vector of uncertainty. \code{errors_max}
(\code{errors_min}) returns the values plus (minus) the uncertainty.
}
\description{
Set or retrieve uncertainty to/from numeric vectors.
}
\details{
\code{`errors<-`} sets the uncertainty values (and converts \code{x}
into an object of class \code{errors}). \code{set_errors} is a pipe-friendly
version of \code{`errors<-`} and returns an object of class \code{errors}.
\code{as.errors} is an alias for \code{set_errors}.

See \code{\link{correl}} on how to handle correlations between pairs of variables.
}
\examples{
x = 1:3
class(x)
x
errors(x) <- 0.1
class(x)
x

(x <- set_errors(x, seq(0.1, 0.3, 0.1)))
errors_max(x)
errors_min(x)

}
\seealso{
\code{\link{groupGeneric.errors}}, \code{\link{mean.errors}},
\code{\link{Extract.errors}}, \code{\link{c}}, \code{\link{rep}}, \code{\link{cbind.errors}},
\code{\link{format.errors}}, \code{\link{print.errors}}, \code{\link{plot.errors}},
\code{\link{as.data.frame.errors}}, \code{\link{as.matrix.errors}}, \code{\link{t}}.
}
