% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.errors}
\alias{format.errors}
\title{Encode \code{errors}}
\usage{
\method{format}{errors}(x, digits = NULL, scientific = FALSE,
  notation = getOption("errors.notation", "parenthesis"), ...)
}
\arguments{
\item{x}{an \code{errors} object.}

\item{digits}{how many significant digits are to be used for uncertainties.
The default, \code{NULL}, uses \code{getOption("errors.digits", 1)}.}

\item{scientific}{logical specifying whether the elements should be
encoded in scientific format.}

\item{notation}{error notation; \code{"parenthesis"} and \code{"plus-minus"}
are supported through the \code{"errors.notation"} option.}

\item{...}{ignored.}
}
\description{
Format an \code{errors} object for pretty printing.
}
\examples{
x <- set_errors(1:3*100, 1:3*100 * 0.05)
format(x)
format(x, digits=2)
format(x, scientific=TRUE)
format(x, notation="plus-minus")

}
