% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.errors}
\alias{format.errors}
\title{Encode errors}
\usage{
\method{format}{errors}(x, digits = NULL, scientific = FALSE,
  notation = getOption("errors.notation", "parenthesis"), ...)
}
\arguments{
\item{x}{an \code{errors} object.}

\item{digits}{how many significant digits are to be used for errors. The default,
\code{NULL}, uses \code{getOption("errors.digits", 1)}.}

\item{scientific}{logical specifying whether the elements should be
encoded in scientific format.}

\item{notation}{error notation; \code{"parenthesis"} and \code{"plus-minus"}
are supported through the \code{"errors.notation"} option.}

\item{...}{ignored.}
}
\description{
Format errors for pretty printing.
}
