% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{mean.errors}
\alias{mean.errors}
\alias{mean.errors}
\alias{weighted.mean.errors}
\alias{mean.errors}
\alias{median.errors}
\title{Arithmetic Mean and Median Value}
\usage{
\method{mean}{errors}(x, ...)

\method{weighted.mean}{errors}(x, ...)

\method{median}{errors}(x, ...)
}
\arguments{
\item{x}{an \code{errors} object.}

\item{...}{further arguments passed to of from other methods.}
}
\value{
An \code{errors} object.
}
\description{
S3 methods for \code{errors} objects.
}
\details{
The \code{mean} and \code{weighted.mean} methods set the error as
the maximum of the standard error of the mean and the (weighted) mean of the errors.

The \code{median} method sets the error as \code{1.253 * errors(mean(x))},
which is derived from the asymptotic variance formula of the median. Note that
this value is valid only if the sample is big enough.
}
