% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.errors}
\alias{plot.errors}
\title{Scatterplot with Error Bars}
\usage{
\method{plot}{errors}(x, y, ...)
}
\arguments{
\item{x}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}

\item{y}{the \code{x} and \code{y} arguments provide the x and y
    coordinates for the plot.  Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link{xy.coords}}
    for details.  If supplied separately, they must be of the same length.}

\item{...}{additional arguments (see \code{\link{plot}}).}
}
\description{
S3 method for \code{errors} objects which automatically prints the error bars.
}
\examples{
cars <- as.matrix(cars)
cars <- as.data.frame(set_errors(cars, cars * 0.05))
plot(cars)

}
