% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.errors}
\alias{plot.errors}
\title{Generic X-Y Plotting}
\usage{
\method{plot}{errors}(x, y, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}

\item{...}{additional arguments (see \code{\link[graphics]{plot}}).}
}
\description{
S3 method for \code{errors} objects which automatically prints the error bars.
}
\examples{
cars <- as.matrix(cars)
cars <- as.data.frame(set_errors(cars, cars * 0.05))
plot(cars)

}
