% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{type_sum.errors}
\alias{type_sum.errors}
\title{\code{type_sum} for Tidy \code{tibble} Printing}
\usage{
type_sum.errors(x, ...)
}
\arguments{
\item{x}{object of class errors}

\item{...}{ignored}
}
\description{
S3 method for \code{errors} objects.
}
